Feat: Add dropdown behavior style
This style requieres a label and an input type checkbox as children of
a container including the dropdown class. Both, label and input elements
are hidden by default.

A media query is needed to display label and input elements while
hidding subcontainer that includes a dropdown-content class and other
elements to control displaying.

To control the last situation, toggle checkbox by using the label.

modified:   src/styles.css