Ref: Update DOM as part of Counter class
Repetitive code to update counter value displayed in DOM across all
event listeners in controls.js. So, remove this paece of code and add it
to Counter class as a method called updateDom(). Also, add a reference
to the HTML element (id equals to #counter) as a new protected property
called itemDom in the mentioned class which is used in the updateDom
method.

modified:   src/js/controls.js
modified:   src/js/counter.js