Feat: Add modal window when click on save button
The Modal window help to display an input field for the description of
the counter to be saved and buttons to accept or cancel this process.

This component include styles to appear in hte middle of screen by using
the grid properties because this component is expected to work in
conjuntion with the background-disabled component, which is a grid
container. The modal is thought as a flex container to position
mentioned elements.

Finally, import this component in toolbar module and append it to the
background-disabled if modal is not previously appended when clicking on
save button.

new file:   src/components/modal-save/modal-save.css
new file:   src/components/modal-save/modal-save.js
modified:   src/js/toolbar.js